backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow-cpu
librosa
noisereduce
soundfile
joblib
numpy
scikit-learn==1.6.1
firebase-admin
